+++ docs/custom.md
@@ -133,9 +133,9 @@ The `TwigCsFixer\Token\Tokenizer` transform the file into a list of tokens which
 
   The `#}` delimiter.
 
-- **TwigCsFixer\Token\Token::NAMED_ARGUMENT_OPERATOR_TYPE**:
+- **TwigCsFixer\Token\Token::NAMED_ARGUMENT_SEPARATOR_TYPE**:
 
-  The `=` or `:` operator used when using named argument. Like `{{ foo(bar=true, baz: false) }}`.
+  The `=` or `:` separator used when using named argument. Like `{{ foo(bar=true, baz: false) }}`.
 
 ### Rule
