+++ src/Rules/Function/NamedArgumentSeparatorRule.php
@@ -2,10 +2,8 @@
 
 declare(strict_types=1);
 
-namespace TwigCsFixer\Rules\Operator;
+namespace TwigCsFixer\Rules\Function;
 
-use Composer\InstalledVersions;
-use Composer\Semver\VersionParser;
 use TwigCsFixer\Environment\StubbedEnvironment;
 use TwigCsFixer\Rules\AbstractFixableRule;
 use TwigCsFixer\Token\Token;
@@ -14,7 +12,7 @@ use TwigCsFixer\Token\Tokens;
 /**
  * Ensures named arguments use `:` syntax instead of `=` (For `twig/twig >= 3.12.0`).
  */
-final class NamedArgumentOperatorRule extends AbstractFixableRule
+final class NamedArgumentSeparatorRule extends AbstractFixableRule
 {
     public function __construct()
     {
@@ -26,7 +24,7 @@ final class NamedArgumentOperatorRule extends AbstractFixableRule
     protected function process(int $tokenIndex, Tokens $tokens): void
     {
         $token = $tokens->get($tokenIndex);
-        if (!$token->isMatching(Token::NAMED_ARGUMENT_OPERATOR_TYPE)) {
+        if (!$token->isMatching(Token::NAMED_ARGUMENT_SEPARATOR_TYPE)) {
             return;
         }
 
@@ -35,7 +33,7 @@ final class NamedArgumentOperatorRule extends AbstractFixableRule
         }
 
         $fixer = $this->addFixableError(
-            \sprintf('Named arguments should be declared with the operator "%s".', ':'),
+            \sprintf('Named arguments should be declared with the separator "%s".', ':'),
             $token
         );
