+++ tests/Rules/Function/NamedArgumentSpacing/NamedArgumentSpacingRuleTest.fixed.twig
