+++ tests/Standard/TwigTest.php
@@ -6,7 +6,7 @@ namespace TwigCsFixer\Tests\Standard;
 
 use PHPUnit\Framework\TestCase;
 use TwigCsFixer\Rules\Delimiter\DelimiterSpacingRule;
-use TwigCsFixer\Rules\Operator\NamedArgumentSpacingRule;
+use TwigCsFixer\Rules\Function\NamedArgumentSpacingRule;
 use TwigCsFixer\Rules\Operator\OperatorNameSpacingRule;
 use TwigCsFixer\Rules\Operator\OperatorSpacingRule;
 use TwigCsFixer\Rules\Punctuation\PunctuationSpacingRule;
