+++ tests/Token/Tokenizer/TokenizerTest.php
@@ -707,7 +707,7 @@ final class TokenizerTest extends TestCase
                 12 => Token::FUNCTION_NAME_TYPE,
                 13 => Token::PUNCTUATION_TYPE,
                 14 => Token::NAME_TYPE,
-                15 => Token::NAMED_ARGUMENT_OPERATOR_TYPE,
+                15 => Token::NAMED_ARGUMENT_SEPARATOR_TYPE,
                 16 => Token::NUMBER_TYPE,
                 17 => Token::PUNCTUATION_TYPE,
                 18 => Token::WHITESPACE_TYPE,
@@ -718,7 +718,7 @@ final class TokenizerTest extends TestCase
                 23 => Token::FUNCTION_NAME_TYPE,
                 24 => Token::PUNCTUATION_TYPE,
                 25 => Token::NAME_TYPE,
-                26 => Token::NAMED_ARGUMENT_OPERATOR_TYPE,
+                26 => Token::NAMED_ARGUMENT_SEPARATOR_TYPE,
                 27 => Token::NUMBER_TYPE,
                 28 => Token::PUNCTUATION_TYPE,
                 29 => Token::WHITESPACE_TYPE,
